fixing translation bugs bringing in fixes from freebsd

+++ src/gamebigmap.cpp
@@ -1,5 +1,5 @@
 //  Copyright (C) 2007, Ole Laursen
-//  Copyright (C) 2007, 2008 Ben Asselstine
+//  Copyright (C) 2007, 2008, 2009 Ben Asselstine
 //
 //  This program is free software; you can redistribute it and/or modify
 //  it under the terms of the GNU General Public License as published by
@@ -455,7 +455,7 @@ void GameBigMap::zoom_out()
       if (new_view.w <= GameMap::getWidth() && 
 	  new_view.h <= GameMap::getHeight() && 
 	  new_view.w >= 0 && new_view.h >= 0)
-	zoom_view(-zoom_step);
+	zoom_view(-(const double)zoom_step);
     }
 }
